server/requirements.txt
flask==3.1.0
flask-cors==5.0.1
dedalus_labs
python-dotenv==1.0.1
